requirements.txt
whisperx==3.3.4
python_dotenv==1.1.0